refactor: kafkaConfig 병합
